neumes now merging correctly, lots of comments added
